requirements.txt
Django
django-angular
django-auth-ldap
django-compressor
django-environ
djangorestframework
djangorestframework-jwt